+++ clients/derivativestradingusdsfutures/tests/unit/restapi/api_trade_test.go
@@ -110,7 +110,6 @@ func Test_binancederivativestradingusdsfuturesrestapi_TradeAPIService(t *testing
 		}
 		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			require.Equal(t, "/fapi/v1/allOrders", r.URL.Path)
-			require.Equal(t, "BTCUSDT", r.URL.Query().Get("symbol"))
 			w.Header().Set("Content-Type", "application/json")
 			_, _ = w.Write([]byte(mockedJSON))
 		}))
@@ -127,7 +126,7 @@ func Test_binancederivativestradingusdsfuturesrestapi_TradeAPIService(t *testing
 			client.WithRestAPI(configuration),
 		)
 
-		resp, err := apiClient.RestApi.TradeAPI.AllOrders(context.Background()).Symbol("BTCUSDT").Execute()
+		resp, err := apiClient.RestApi.TradeAPI.AllOrders(context.Background()).Execute()
 		require.NoError(t, err)
 		require.NotNil(t, resp)
 		require.Equal(
@@ -138,23 +137,6 @@ func Test_binancederivativestradingusdsfuturesrestapi_TradeAPIService(t *testing
 		require.Equal(t, reflect.TypeOf(models.AllOrdersResponse{}), reflect.TypeOf(resp.Data))
 		require.Equal(t, 200, resp.Status)
 		require.Equal(t, expected, resp.Data)
-	})
-
-	t.Run("Test TradeAPIService AllOrders Missing Required Params", func(t *testing.T) {
-		mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
-		defer mockServer.Close()
-
-		configuration := common.NewConfigurationRestAPI()
-		configuration.BasePath = mockServer.URL
-
-		apiClient := client.NewBinanceDerivativesTradingUsdsFuturesClient(
-			client.WithRestAPI(configuration),
-		)
-
-		resp, err := apiClient.RestApi.TradeAPI.AllOrders(context.Background()).Execute()
-
-		require.Error(t, err)
-		require.Nil(t, resp)
 	})
 
 	t.Run("Test TradeAPIService AllOrders Server Error", func(t *testing.T) {